# src/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Eventos API",
    "description": "API para gerenciamento de eventos"
  },
  "basePath": "/",
  "paths": {
    "/email/{id}": {
      "post": {
        "summary": "Enviar e-mail para um usuário",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID do usuário",
            "required": true,
            "type": "integer"
          },
          {
            "name": "body",
            "in": "body",
            "description": "Objeto de e-mail a ser enviado",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Email"
            }
          },
          {
            "name": "username",
            "in": "header",
            "description": "Credenciais de autenticação(username)",
            "required": true,
            "type": "string"
          },
          {
            "name": "password",
            "in": "header",
            "description": "Credenciais de autenticação(password)",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Operação bem-sucedida"
          }
        },
        "examples": {
          "application/json": {
            "mensagem": "E-mail enviado com sucesso"
          }
        }
      }
    },
    "/eventos": {
      "get": {
        "summary": "Listar todos os eventos",
        "parameters": [
          {
            "name": "username",
            "in": "header",
            "description": "Credenciais de autenticação(username)",
            "required": true,
            "type": "string"
          },
          {
            "name": "password",
            "in": "header",
            "description": "Credenciais de autenticação(password)",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Operação bem-sucedida",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Evento"
              }
            }
          }
        }
      },
      "post": {
        "summary": "Criar um novo evento",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Objeto de evento a ser criado",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Evento"
            }
          },
          {
            "name": "username",
            "in": "header",
            "description": "Credenciais de autenticação(username)",
            "required": true,
            "type": "string"
          },
          {
            "name": "password",
            "in": "header",
            "description": "Credenciais de autenticação(password)",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Operação bem-sucedida"
          }
        }
      }
    },
    "/eventos/{id}": {
      "get": {
        "summary": "Obter detalhes de um evento",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID do evento",
            "required": true,
            "type": "integer"
          },
          {
            "name": "username",
            "in": "header",
            "description": "Credenciais de autenticação(username)",
            "required": true,
            "type": "string"
          },
          {
            "name": "password",
            "in": "header",
            "description": "Credenciais de autenticação(password)",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Operação bem-sucedida",
            "schema": {
              "$ref": "#/definitions/Evento"
            }
          }
        }
      },
      "put": {
        "summary": "Atualizar um evento existente",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID do evento",
            "required": true,
            "type": "integer"
          },
          {
            "name": "body",
            "in": "body",
            "description": "Objeto de evento com as atualizações",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Evento"
            }
          },
          {
            "name": "username",
            "in": "header",
            "description": "Credenciais de autenticação(username)",
            "required": true,
            "type": "string"
          },
          {
            "name": "password",
            "in": "header",
            "description": "Credenciais de autenticação(password)",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Operação bem-sucedida"
          }
        }
      },
      "delete": {
        "summary": "Deletar um evento existente",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID do evento",
            "required": true,
            "type": "integer"
          },
          {
            "name": "username",
            "in": "header",
            "description": "Credenciais de autenticação(username)",
            "required": true,
            "type": "string"
          },
          {
            "name": "password",
            "in": "header",
            "description": "Credenciais de autenticação(password)",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Operação bem-sucedida"
          }
        }
      }
    },
    "/inscricoes": {
      "get": {
        "summary": "Listar todas as inscrições",
        "parameters": [
          {
            "name": "username",
            "in": "header",
            "description": "Credenciais de autenticação(username)",
            "required": true,
            "type": "string"
          },
          {
            "name": "password",
            "in": "header",
            "description": "Credenciais de autenticação(password)",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Operação bem-sucedida",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Inscricao"
              }
            }
          }
        }
      },
      "post": {
        "summary": "Realizar inscrição em um evento",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Objeto de inscrição a ser realizado",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Inscricao"
            }
          },
          {
            "name": "username",
            "in": "header",
            "description": "Credenciais de autenticação(username)",
            "required": true,
            "type": "string"
          },
          {
            "name": "password",
            "in": "header",
            "description": "Credenciais de autenticação(password)",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Operação bem-sucedida"
          }
        }
      }
    },
    "/inscricoes/{id}": {
      "get": {
        "summary": "Obter detalhes de uma inscrição",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID da inscrição",
            "required": true,
            "type": "integer"
          },
          {
            "name": "username",
            "in": "header",
            "description": "Credenciais de autenticação(username)",
            "required": true,
            "type": "string"
          },
          {
            "name": "password",
            "in": "header",
            "description": "Credenciais de autenticação(password)",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Operação bem-sucedida",
            "schema": {
              "$ref": "#/definitions/Inscricao"
            }
          }
        }
      },
      "put": {
        "summary": "Atualizar uma inscrição existente",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID da inscrição",
            "required": true,
            "type": "integer"
          },
          {
            "name": "body",
            "in": "body",
            "description": "Objeto de inscrição com as atualizações",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Inscricao"
            }
          },
          {
            "name": "username",
            "in": "header",
            "description": "Credenciais de autenticação(username)",
            "required": true,
            "type": "string"
          },
          {
            "name": "password",
            "in": "header",
            "description": "Credenciais de autenticação(password)",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Operação bem-sucedida"
          }
        }
      },
      "delete": {
        "summary": "Deletar uma inscrição existente",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID da inscrição",
            "required": true,
            "type": "integer"
          },
          {
            "name": "username",
            "in": "header",
            "description": "Credenciais de autenticação(username)",
            "required": true,
            "type": "string"
          },
          {
            "name": "password",
            "in": "header",
            "description": "Credenciais de autenticação(password)",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Operação bem-sucedida"
          }
        }
      }
    },
    "/usuarios": {
      "get": {
        "summary": "Listar todos os usuários",
        "parameters": [
          {
            "name": "username",
            "in": "header",
            "description": "Credenciais de autenticação(username)",
            "required": true,
            "type": "string"
          },
          {
            "name": "password",
            "in": "header",
            "description": "Credenciais de autenticação(password)",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Operação bem-sucedida",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Usuario"
              }
            }
          }
        }
      },
      "post": {
        "summary": "Criar um novo usuário",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Objeto de usuário a ser criado",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Usuario"
            }
          },
          {
            "name": "username",
            "in": "header",
            "description": "Credenciais de autenticação(username)",
            "required": true,
            "type": "string"
          },
          {
            "name": "password",
            "in": "header",
            "description": "Credenciais de autenticação(password)",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Operação bem-sucedida"
          }
        }
      }
    },
    "/usuarios/{id}": {
      "get": {
        "summary": "Obter detalhes de um usuário",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID do usuário",
            "required": true,
            "type": "integer"
          },
          {
            "name": "username",
            "in": "header",
            "description": "Credenciais de autenticação(username)",
            "required": true,
            "type": "string"
          },
          {
            "name": "password",
            "in": "header",
            "description": "Credenciais de autenticação(password)",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Operação bem-sucedida",
            "schema": {
              "$ref": "#/definitions/Usuario"
            }
          }
        }
      },
      "put": {
        "summary": "Atualizar um usuário existente",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID do usuário",
            "required": true,
            "type": "integer"
          },
          {
            "name": "body",
            "in": "body",
            "description": "Objeto de usuário com as atualizações",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Usuario"
            }
          },
          {
            "name": "username",
            "in": "header",
            "description": "Credenciais de autenticação(username)",
            "required": true,
            "type": "string"
          },
          {
            "name": "password",
            "in": "header",
            "description": "Credenciais de autenticação(password)",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Operação bem-sucedida"
          }
        }
      },
      "delete": {
        "summary": "Deletar um usuário existente",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID do usuário",
            "required": true,
            "type": "integer"
          },
          {
            "name": "username",
            "in": "header",
            "description": "Credenciais de autenticação(username)",
            "required": true,
            "type": "string"
          },
          {
            "name": "password",
            "in": "header",
            "description": "Credenciais de autenticação(password)",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Operação bem-sucedida"
          }
        }
      }
    }
  },
  "definitions": {
    "Evento": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer"
        },
        "nome": {
          "type": "string"
        },
        "dataInicio": {
          "type": "string"
        },
        "dataFim": {
          "type": "string"
        },
        "hora": {
          "type": "string"
        },
        "local": {
          "type": "string"
        },
        "descricao": {
          "type": "string"
        }
      }
    },
    "Inscricao": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer"
        },
        "status": {
          "type": "boolean"
        },
        "checkin": {
          "type": "boolean"
        }
      }
    },
    "Usuario": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer"
        },
        "nome": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "senha": {
          "type": "string"
        },
        "telefone": {
          "type": "string"
        }
      }
    },
    "Email": {
      "type": "object",
      "properties": {
        "titulo": {
          "type": "string"
        },
        "mensagem": {
          "type": "string"
        }
      }
    }
  }
}